Añadir test read reports y modificació eliminació dades base de dades local

+++ back_api/tests/setup.js
@@ -1,6 +1,7 @@
 const firebase = require('@firebase/testing');
+require('dotenv').config();
 
-const projectId = 'culturapp-82c6c'; // Reemplaza esto con el ID de tu proyecto
+const projectId = `${process.env.PROJECT_ID}`;
 
 beforeAll(async () => {
   console.log('Initializing Firestore');
@@ -8,7 +9,7 @@ beforeAll(async () => {
   db = app.firestore();
 });
 
-afterAll(async () => {
+afterEach(async () => {
   console.log('Deleting all documents');
   
   const collections = ['reportsUsuaris']; // Agregar aquí todas las colecciones que se quieran borrar
@@ -16,10 +17,16 @@ afterAll(async () => {
   for (const collection of collections) {
     const snapshot = await db.collection(collection).get();
 
+    if(snapshot.empty) continue;    
+
     // Borrar todos los documentos en la colección
     const deletePromises = snapshot.docs.map(doc => db.collection(collection).doc(doc.id).delete());
     await Promise.all(deletePromises);
   }
+});
+
+afterAll(async () => {
   console.log('Closing Firestore');
   await Promise.all(firebase.apps().map(app => app.delete()));
 });
+
